estateProject/requirements.txt
amqp==5.3.1
anyio==4.11.0
arabic-reshaper==3.0.0
asgiref==3.10.0
asn1crypto==1.5.1
attrs==25.4.0
autobahn==25.10.2
Automat==25.4.16
billiard==4.2.2
Brotli==1.1.0
CacheControl==0.14.3
cachetools==6.2.1
celery==5.4.0
certifi==2025.1.31
cffi==2.0.0
channels==4.3.1
channels_redis==4.3.0
chardet==5.2.0
charset-normalizer==3.4.1
click==8.3.0
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
constantly==23.10.4
cryptography==46.0.2
cssselect2==0.8.0
daphne==4.2.1
dj-database-url==3.0.1
Django==5.2.8
django-cors-headers==4.7.0
django-filter==25.2
django-model-utils==5.0.0
django-multiselectfield==0.1.13
django-rest-framework==0.1.0
django-widget-tweaks==1.5.0
djangorestframework==3.15.2
drf-spectacular==0.29.0
et_xmlfile==2.0.0
firebase-admin==6.5.0
fonttools==4.60.1
freetype-py==2.5.1
git-filter-repo==2.47.0
google-api-core==2.28.1
google-api-python-client==2.186.0
google-auth==2.42.1
google-auth-httplib2==0.2.1
google-cloud-core==2.5.0
google-cloud-firestore==2.21.0
google-cloud-storage==3.4.1
google-crc32c==1.7.1
google-resumable-media==2.7.2
googleapis-common-protos==1.71.0
grpcio==1.76.0
grpcio-status==1.76.0
gunicorn==23.0.0
h11==0.16.0
h2==4.3.0
hpack==4.1.0
html5lib==1.1
httpcore==1.0.9
httplib2==0.31.0
httpx==0.28.1
hyperframe==6.1.0
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
inflection==0.5.1
jsonfield==3.1.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
kombu==5.4.2
lxml==6.0.2
msgpack==1.1.2
openpyxl==3.1.5
oscrypto==1.3.0
packaging==25.0
pillow==11.1.0
prompt_toolkit==3.0.52
proto-plus==1.26.1
protobuf==6.33.0
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycairo==1.28.0
pycparser==2.23
pydyf==0.11.0
pyHanko==0.31.0
pyhanko-certvalidator==0.29.0
PyJWT==2.10.1
pyOpenSSL==25.3.0
pyparsing==3.2.5
pypdf==6.1.1
pyphen==0.17.2
python-bidi==0.6.6
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.1
PyYAML==6.0.3
redis==7.0.1
referencing==0.37.0
reportlab==4.4.4
requests==2.32.3
rlPyCairo==0.4.0
rpds-py==0.29.0
rsa==4.9.1
sentry-sdk==2.45.0
service-identity==24.2.0
setuptools==75.8.2
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.3
stripe==14.0.0
svglib==1.6.0
swapper==1.4.0
tinycss2==1.4.0
tinyhtml5==2.0.0
Twisted==25.5.0
txaio==25.9.2
typing_extensions==4.15.0
tzdata==2024.2
tzlocal==5.3.1
uritemplate==4.2.0
uritools==5.0.0
urllib3==2.3.0
vine==5.1.0
wcwidth==0.2.14
weasyprint==66.0
webencodings==0.5.1
whitenoise==6.11.0
xhtml2pdf==0.2.17
zope.interface==8.0.1
zopfli==0.2.3.post1
